refactor: migrate persistence stores into domains
